--- .env
@@ -1,14 +1,0 @@
-HTTPS=true
-SSL_CRT_FILE=./src/localhost.pem
-SSL_KEY_FILE=./src/localhost-key.pem
-
-
-# This was inserted by `prisma init`:
-# Environment variables declared in this file are automatically made available to Prisma.
-# See the documentation for more detail: https://pris.ly/d/prisma-schema#accessing-environment-variables-from-the-schema
-
-# Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB and CockroachDB.
-# See the documentation for all the connection string options: https://pris.ly/d/connection-strings
-SESSION_SECRET=secret
-DATABASE_URL=postgresql://kwaai:secret@localhost:5432/kwaai_db?schema=public
-PAIOS_API_URL=http://localhost:3080/api/v1
