--- docker-compose.yml
@@ -1,13 +1,0 @@
-version: '3.9'
-
-services:
-  postgres:
-    image: postgres:14-alpine
-    ports:
-      - 5432:5432
-    # volumes:
-    #   - ~/apps/postgres:/var/lib/postgresql/data
-    environment:
-      - POSTGRES_PASSWORD=secret
-      - POSTGRES_USER=kwaai
-      - POSTGRES_DB=kwaai_db
