--- prisma/schema.prisma
@@ -1,33 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-// Looking for ways to speed up your queries, or scale easily with your serverless or edge functions?
-// Try Prisma Accelerate: https://pris.ly/cli/accelerate-init
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-model UserCredential {
-  id       String @id
-  raw_id   String
-  response Json
-  type     String
-  email    String @unique
-}
-
-model User {
-  email String @unique
-  id    String @id
-}
-
-model Session {
-  sid     String   @id
-  sess    Json
-  expire  DateTime 
-}
-
-datasource db {
-  provider = "postgresql"
-  url      = env("DATABASE_URL")
-}
