+++ public/index.html
@@ -2,14 +2,12 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="/src/assets/kwaai.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
       content="Web site created using create-react-app"
     />
-    <link rel="apple-touch-icon" href="/src/assets/kwaai.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <title>Kwaai pAI</title>
   </head>
