--- src/components/PrivateRoute.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const PrivateRoute = ({ children }: { children: JSX.Element }) => {
-  return children;
-};
-
-export default PrivateRoute;
