--- src/components/profileDropdown/profileDropdown.css
@@ -1,47 +1,0 @@
-.profile-dropdown {
-    position: absolute; /* Change to absolute positioning */
-    top: 10px; /* Adjust the top position as needed */
-    right: 50px; /* Position it to the right */
-    display: inline-block;
-  }
-  
-  .profile-button {
-    background-color: transparent; /* Make the background transparent */
-    color: black; /* Change text color to black */
-    padding: 10px;
-    font-size: 16px;
-    border: none;
-    cursor: pointer;
-    border-radius: 50%; /* Make the button a circle */
-    width: auto; /* Adjust width to fit content */
-    height: 40px; /* Set a fixed height */
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0 15px; /* Add padding for better appearance */
-  }
-  
-  .dropdown-menu {
-    display: block;
-    position: absolute;
-    background-color: white;
-    min-width: 160px;
-    box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
-    z-index: 1;
-  }
-  
-  .dropdown-menu button {
-    color: black;
-    padding: 12px 16px;
-    text-decoration: none;
-    display: block;
-    width: 100%;
-    border: none;
-    background: none;
-    text-align: left;
-    cursor: pointer;
-  }
-  
-  .dropdown-menu button:hover {
-    background-color: #ddd;
-  }
