--- src/localhost-key.pem
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCadvByrbXYn6/J
-BfFLGqm6hgy11rMgwFAgjdjdOtzkFGiiuoYUZcYxSv00/9u572G1WENopx1FAfSt
-Pa1DjtUcbITWbbBYgT5m4T/dQIW7aH8rqTyYtOsE+M52Z36MqvcpqBLzLtrQJE4h
-5+HzapojIgsh/eWGvwnSJNxG+GeWRMAZ8tAvZVgreLXGB5CzxaEbjh69fMnmTZHQ
-BSjWuXnYZ9vUqgu87cD8lFIBxRfjmsx47eExG/3KCHLCJWsR1mGdgOOQ6VwSe7C8
-X5pklvz0fhcJTjAmBSqHEcOTKLeQnWnegOslEWUKYGa1ZJZ/60H7xPZECBT7jsws
-sZAKikDTAgMBAAECggEAGJxez0bSNIOs9HCX84l6BZgaKlbdD5jUq3suVIFudICz
-80mY12Mn4u65zVn+3m8/uC8p/mCyDy6yl2++TYdSeUbpG2k6sb9V+/pjvJLtO+D4
-cjJyzzn0QESMUN1Iko0i+GYpLsoGXyeJs2lI3QndKT/hWjcj/QYvQGZTGpO1nYFz
-X/b70psZNxZ5eVg/lsU/ar7PgPcpEqEGqNWtQYBPZ6bXjq3W9O3CPyMnlq2p1OkJ
-+UKSPPIpvweiXPOj9ZGMWr58QSnX5KnWdknPo3A1+UWVcLTYyFbJtxKDkZI5olEG
-cBT0IqwvIdvVhimTi+bNBmTxrOFl2rQJXGkZGcxOEQKBgQDCcgXaMF544gWn7tzN
-wzsvc2aUXj+4HPAmuRY6r2bkLugYuwWLYwG50bscVfZOpInr+GIRLAiNbC0VGwnT
-/qYRzqOWcjdRdeOMCeIQ5DERzXLD5eK92mVcj0KrUX5h+6K3tz9YEcLMHKRwY+xJ
-FJZaQ5iYxFpDrqWDTUGq2HOdDQKBgQDLXNkCgnYIqn4rtLQtaezlqyPO08XrQIMX
-yU5lFAOrMDwqU2r+sJgnNnvh7WIE2yo0iXTtNMZkOe5bPF0Gynw3YfEvHtcVlSXe
-k8oHrmUAWOvglzpgkMVcwTraYAWOAV/hinbeJ1qEYgQnTVL+4bAuccNV2pvBaU3M
-26Z7ERqdXwKBgQCqq9KfNR13FUzJnkuI//gPKke7qw5a2GuiUYtFb3fsiY+HRYEU
-W47EP2bw9DMBzEKTMK1DTVckavMXN2hk/3tc3feQsigEBGn6HSKkmObJLX+ymzFu
-WyflGon2Pc2ky26DuwCKmOxWpD1huRNJAh0GlZZ3YMGp7r2whjtBbkPscQKBgQCg
-UuOW8elCLb9oj37YqsCj1AdX4GcRb3uqPf86+stlo5JbQAaDIg/iIjcKR45WC767
-Sn7PHyMlEvLSJMWmO+4v71wobOTeVAZV8id9ZMSlwHS9TBwu0xN1a7UNQlQ+vuCr
-O/E4dMgMxAVIQtKitKSkCPYZck4G7hDy51vZfAOCewKBgD++N3+vc4cZkCVuP3fi
-tRO5K3/qSgvWZ5xiSG7JzhTVzrB/1B8fCp/2PiwpHm8YvH4FAKDJ7bU3+yQ3fDFU
-p13zH8c/hDHUChVHIa4etYm55zK51GPOY7ZXt84bpoSjXDiT1+xQxI/U9DYJ0F1+
-S2E3TdgpJBID85novzEsMPkC
------END PRIVATE KEY-----
