--- src/localhost.pem
@@ -1,27 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIEjzCCAvegAwIBAgIRANGReBX5t08AnH5aou4F4SMwDQYJKoZIhvcNAQELBQAw
-gbUxHjAcBgNVBAoTFW1rY2VydCBkZXZlbG9wbWVudCBDQTFFMEMGA1UECww8QXp1
-cmVBRFxFbWlsaWFub1PDoW5jaGV6T2xndWlARGVsbEk3IChFbWlsaWFub1PDoW5j
-aGV6T2xndWkpMUwwSgYDVQQDDENta2NlcnQgQXp1cmVBRFxFbWlsaWFub1PDoW5j
-aGV6T2xndWlARGVsbEk3IChFbWlsaWFub1PDoW5jaGV6T2xndWkpMB4XDTI0MDgx
-NjE3MjkzM1oXDTI2MTExNjE3MjkzM1owcDEnMCUGA1UEChMebWtjZXJ0IGRldmVs
-b3BtZW50IGNlcnRpZmljYXRlMUUwQwYDVQQLDDxBenVyZUFEXEVtaWxpYW5vU8Oh
-bmNoZXpPbGd1aUBEZWxsSTcgKEVtaWxpYW5vU8OhbmNoZXpPbGd1aSkwggEiMA0G
-CSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCadvByrbXYn6/JBfFLGqm6hgy11rMg
-wFAgjdjdOtzkFGiiuoYUZcYxSv00/9u572G1WENopx1FAfStPa1DjtUcbITWbbBY
-gT5m4T/dQIW7aH8rqTyYtOsE+M52Z36MqvcpqBLzLtrQJE4h5+HzapojIgsh/eWG
-vwnSJNxG+GeWRMAZ8tAvZVgreLXGB5CzxaEbjh69fMnmTZHQBSjWuXnYZ9vUqgu8
-7cD8lFIBxRfjmsx47eExG/3KCHLCJWsR1mGdgOOQ6VwSe7C8X5pklvz0fhcJTjAm
-BSqHEcOTKLeQnWnegOslEWUKYGa1ZJZ/60H7xPZECBT7jswssZAKikDTAgMBAAGj
-XjBcMA4GA1UdDwEB/wQEAwIFoDATBgNVHSUEDDAKBggrBgEFBQcDATAfBgNVHSME
-GDAWgBQ6CGv4Cgt2AqLd43YM/oaEiQPqTzAUBgNVHREEDTALgglsb2NhbGhvc3Qw
-DQYJKoZIhvcNAQELBQADggGBAIWw/Z8f/tVw01iZi/PEQs8yMUZu133Ap6NJ1gFO
-1fdyeHv4fhakm5ais4UDbGXLK+Kf5zPQtR5c5i/PjBjVXe9DbOjQscDNIzXXEbS7
-sRht2d+BLXAzNfH31mhYRz1I3fNbTqbTMxBHpPZckCjPXKnnkzV00c6gEBN/Nq+X
-to9f8KQy1pAcyKvmh1ZiJ6IF18N+ebM7ukSEjUPRaMg8L5+iWqGfZJd/IS6IQByG
-9FgTM5ILbbzEoiiq0PghWodSQJk0gyOZTsV9DMtOuMAVUP4IsPgMODoXRrKwXX0G
-YPayilh2+W4H8g1ZWgGoAcP6abkjA6rG73wHmOfmiZMkahkjF3doqAk5MWBu7TvO
-FWeQAX3Jqqo/kowSS6bgzgMoHZPJoL3MIyPp4YeQIGBtBM/I/EEFRwJ3zlYDcGO/
-sBrBNgYBHhx/FEvaR35+IyENrbNHfQtm90Ic/NPYX/g72JLQlR3TRN+IV32Mmt0I
-T2nWwMhzu1bTkgZdWlDWdtfjPA==
------END CERTIFICATE-----
