--- src/pages/register/Register.css
@@ -1,45 +1,0 @@
-.auth-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 100vh;
-    background-color: #f5f5f5;
-}
-
-.logo {
-    width: 150px;
-    margin-bottom: 20px;
-}
-
-h1 {
-    font-size: 2em;
-    margin-bottom: 20px;
-}
-
-.input-field {
-    width: 300px;
-    padding: 10px;
-    margin-bottom: 20px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-}
-
-.auth-button {
-    width: 300px;
-    padding: 10px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-.auth-link {
-    margin-top: 20px;
-}
-
-.auth-link a {
-    color: #007bff;
-    text-decoration: none;
-}
