--- src/users.js
@@ -1,32 +1,0 @@
-import * as React from 'react';
-import { List, Datagrid, TextField, EditButton, Edit, SimpleForm, TextInput, Create } from 'react-admin';
-
-export const UserList = (props) => (
-    <List {...props}>
-        <Datagrid>
-            <TextField source="id" />
-            <TextField source="name" />
-            <TextField source="email" />
-            <EditButton />
-        </Datagrid>
-    </List>
-);
-
-export const UserEdit = (props) => (
-    <Edit {...props}>
-        <SimpleForm>
-            <TextInput disabled source="id" />
-            <TextInput source="name" />
-            <TextInput source="email" />
-        </SimpleForm>
-    </Edit>
-);
-
-export const UserCreate = (props) => (
-    <Create {...props}>
-        <SimpleForm>
-            <TextInput source="name" />
-            <TextInput source="email" />
-        </SimpleForm>
-    </Create>
-);
